Clean up code & expand mock data
