+++ data/fieldData.py
@@ -1,0 +1,11 @@
+fieldData = [
+  {
+    "name": "wikidata_qid", 
+    "description": "a thing for doing the thing"
+  },
+  {
+    "name": "tool_type", 
+    "description": "this other thing",
+    "input_options": "a, b, c"
+  }
+]
