+++ data/taskData.py
@@ -1,0 +1,1 @@
+taskData = [ { "tool_name": "mm_wikidata_todo", "field_name": "wikidata_qid" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type", "user": "Demo Unicorn" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type", "user": "Demo Unicorn" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type", "user": "Demo Unicorn" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type", "user": "Demo Unicorn" }, { "tool_name": "mm_wikidata_todo", "field_name": "tool_type", "user": "NicoleLBee" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "Hannah Waruguru" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "Hannah Waruguru" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "Hannah Waruguru" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "BDavis (WMF)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "Sage (Wiki Ed)" }, { "tool_name": "pywikibot", "field_name": "tool_type", "user": "Sage (Wiki Ed)" }, { "tool_name": "pywikibot", "field_name": "wikidata_qid", "user": "Sage (Wiki Ed)" }, { "tool_name": "pywikibot", "field_name": "wikidata_qid", "user": "Sage (Wiki Ed)" }, { "tool_name": "pywikibot", "field_name": "wikidata_qid", "user": "Sage (Wiki Ed)" }, { "tool_name": "mm_wikidata_todo", "field_name": "wikidata_qid" } ]
