+++ data/toolData.py
@@ -1,0 +1,15 @@
+toolData = [
+  {
+    "title": "Wikidata Todo",
+    "name": "mm_wikidata_todo",
+    "description": "Shows you little things you can do on Wikidata.",
+    "url": "http://tools.wmflabs.org/wikidata-todo"
+  },
+  {
+    "title": "Pywikibot",
+    "name": "pywikibot",
+    "description":
+      "Python library and collection of scripts that automate work on MediaWiki sites",
+    "url": "https://www.mediawiki.org/wiki/Special:MyLanguage/Manual:Pywikibot"
+  }
+]
